requirements.txt
python-pptx==0.6.23
PyPDF2==3.0.1
openai==1.12.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
python-docx==0.8.11